Move UpperCaseCharField to Field file
