+++ dakara_server/library/fields.py
@@ -12,3 +12,21 @@ class SafeDurationField(models.DurationField):
         if value is None:
             return None
         return int(round(value))
+
+
+class UpperCaseCharField(models.CharField):
+    """ Override a Django Model Field and make it upper-case as of Django 1.8
+        http://stackoverflow.com/a/33354171
+    """
+
+    def __init__(self, *args, **kwargs):
+        super(UpperCaseCharField, self).__init__(*args, **kwargs)
+
+    def pre_save(self, model_instance, add):
+        value = getattr(model_instance, self.attname, None)
+        if value:
+            value = value.upper()
+            setattr(model_instance, self.attname, value)
+            return value
+        else:
+            return super(UpperCaseCharField, self).pre_save(model_instance, add)
